Add files via upload
Focus on five learning objectives:

 1. Define the central concepts of encryption and decryption, preview a project demo, and explore its functionality to understand what I aim to build by the end of this project.  
 2. Engage in hands-on learning about ASCII codes and typecasting.  
 3. Participate in practical learning regarding the Vigenère cypher and apply this concept to encrypt a single character.  
 4. Undertake hands-on learning on encrypting an entire string using the Vigenère cypher.  
 5. Create a basic application for encryption and decryption utilising the Vigenère cypher.  

By the end of this course, I was equipped to identify the fundamentals of cryptography and learn how to encrypt and decrypt messages using a specific technique. I successfully applied concepts such as ASCII codes and typecasting. Additionally, I performed mathematical operations on characters, ultimately preparing me to delve deeper into the world of cryptography.

Signed-off-by: Gowtham T <174505325+gowthamthotakura@users.noreply.github.com>